smol changes. got some movement fixed, but still have more to do with respect to out of bounds
